[refactor/FIN-69] 추가 정보 등록 API 기반 리팩토링
- User, UsersBlog 로 분리된 도메인 병합
- User Entity 변경
- 기존의 UsersBlogApi, UsersBlogService, UsersBlogRepository 모두 삭제
- User의 Api, Service, Repository가 삭제된 부분 모두 구현
- src/blog 패키지를 src/article 로 변경
