+++ src/main/java/kr/co/finote/backend/src/article/repository/ArticleRepository.java
@@ -1,7 +1,7 @@
-package kr.co.finote.backend.src.blog.repository;
+package kr.co.finote.backend.src.article.repository;
 
 import java.util.Optional;
-import kr.co.finote.backend.src.blog.domain.Article;
+import kr.co.finote.backend.src.article.domain.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
