+++ src/main/java/kr/co/finote/backend/src/article/service/ArticleService.java
@@ -1,10 +1,10 @@
-package kr.co.finote.backend.src.blog.service;
+package kr.co.finote.backend.src.article.service;
 
 import kr.co.finote.backend.global.code.ResponseCode;
 import kr.co.finote.backend.global.exception.CustomException;
-import kr.co.finote.backend.src.blog.domain.Article;
-import kr.co.finote.backend.src.blog.dto.request.ArticleRequest;
-import kr.co.finote.backend.src.blog.repository.ArticleRepository;
+import kr.co.finote.backend.src.article.domain.Article;
+import kr.co.finote.backend.src.article.dto.request.ArticleRequest;
+import kr.co.finote.backend.src.article.repository.ArticleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
